feat: add BasePagination component + use in CardPagination and Table components
