+++ packages/shira-ui/src/components/BasePagination/index.ts
@@ -1,0 +1,1 @@
+export * from "./BasePagination";
