+++ packages/shira-ui/src/components/CardPagination/CardPagination.tsx
@@ -1,8 +1,5 @@
 import { FunctionComponent } from "react";
-import styled, { useTheme } from "styled-components";
-import { darken } from "polished";
-import { FiChevronLeft, FiChevronRight, FiChevronsLeft, FiChevronsRight } from "react-icons/fi";
-import { Body2Regular, Body2SemiBold, Body3 } from "../Typography";
+import { BasePagination } from "../BasePagination";
 
 interface Props {
   pageIndex: number
@@ -25,126 +22,21 @@ export const CardPagination: FunctionComponent<Props> = ({
   onNextPage,
   onLastPage,
 }) => {
-  const theme = useTheme();
   const canPreviousPage = pageIndex > 0;
   const canNextPage = pageIndex < pageCount - 1;
-  const offSetPageIndex = pageIndex + 1;
-
-  const start = total === 0 ? 0 : pageIndex * PAGE_SIZE + 1;
-  const end = Math.min((pageIndex + 1) * PAGE_SIZE, total);
 
   return (
-    <PaginationWrapper>
-      <StyledBody2SemiBold>
-        {`${start}-${end} of ${total}`}
-      </StyledBody2SemiBold>
-      <PaginationButtons>
-        <PaginationButton
-          type="button"
-          disabled={!canPreviousPage}
-          onClick={onFirstPage}
-        >
-          <FiChevronsLeft size={16} color={theme.colors.dark.darkGrey} />
-          <Body3>First</Body3>
-        </PaginationButton>
-        <PaginationButton
-          type="button"
-          disabled={!canPreviousPage}
-          onClick={onPreviousPage}
-        >
-          <FiChevronLeft size={16} color={theme.colors.dark.darkGrey} />
-          <Body3>Back</Body3>
-        </PaginationButton>
-
-        <PaginationNavButton>
-          <StyledBody2Bold>{`${offSetPageIndex} `}</StyledBody2Bold>
-          <Body2>{`of ${pageCount}`}</Body2>
-        </PaginationNavButton>
-
-        <PaginationButton
-          type="button"
-          disabled={!canNextPage}
-          onClick={onNextPage}
-        >
-          <Body3>Next</Body3>
-          <FiChevronRight size={16} color={theme.colors.dark.darkGrey} />
-        </PaginationButton>
-        <PaginationButton
-          type="button"
-          disabled={!canNextPage}
-          onClick={onLastPage}
-        >
-          <Body3>Last</Body3>
-          <FiChevronsRight size={16} color={theme.colors.dark.darkGrey} />
-        </PaginationButton>
-      </PaginationButtons>
-    </PaginationWrapper>
+    <BasePagination
+      pageIndex={pageIndex}
+      pageCount={pageCount}
+      pageSize={PAGE_SIZE}
+      total={total}
+      canPreviousPage={canPreviousPage}
+      canNextPage={canNextPage}
+      onFirstPage={onFirstPage}
+      onPreviousPage={onPreviousPage}
+      onNextPage={onNextPage}
+      onLastPage={onLastPage}
+    />
   );
 };
-
-const StyledBody2Bold = styled(Body2SemiBold)`
-  color: ${props => props.theme.colors.dark.darkGrey};
-  font-weight: 700;
-`;
-
-const StyledBody2SemiBold = styled(Body2SemiBold)`
-  color: ${props => props.theme.colors.dark.darkGrey};
-`;
-
-const Body2 = styled(Body2Regular)`
-  color: ${props => props.theme.colors.dark.darkGrey};
-`;
-
-const PaginationWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 12px 0;
-  width: 100%;
-  margin: 0;
-`;
-
-const PaginationButton = styled.button<{
-  disabled: boolean
-}>`
-  all: unset;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  padding: 6px 10px;
-  color: ${props => props.theme.colors.dark.darkGrey};
-  border-radius: 4px;
-  border: 2px solid white;
-  background: white;
-  gap: 8px;
-
-  ${({ disabled }) => disabled && `
-    opacity: 0.5;
-    cursor: not-allowed;
-  `}
-
-  ${({ disabled, theme }) => !disabled && `
-    &:hover {
-      background: ${darken(0.1, 'white')};
-      border-color: ${darken(0.1, 'white')};
-      color: ${theme.colors.dark.black};
-    }
-    &:active {
-      background: white;
-      border-color: ${theme.colors.green4};
-      color: ${theme.colors.dark.black};
-    }
-  `}
-`;
-
-const PaginationButtons = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-`;
-
-const PaginationNavButton = styled(PaginationButtons)`
-  display: flex;
-  align-items: center;
-  gap: 4px;
-`;
