+++ packages/shira-ui/src/components/Table/components/Pagination.tsx
@@ -1,9 +1,6 @@
 import { Table } from "@tanstack/react-table";
 import { FunctionComponent } from "react";
-import styled, { useTheme } from "styled-components";
-import { Body2Regular, Body2SemiBold, Body3 } from "../../Typography";
-import { FiChevronLeft, FiChevronRight, FiChevronsLeft, FiChevronsRight } from "react-icons/fi";
-import { darken } from "polished";
+import { BasePagination } from "../../BasePagination";
 
 interface Props {
   table: Table<unknown>
@@ -12,124 +9,21 @@ interface Props {
 export const Pagination: FunctionComponent<Props> = ({
   table
 }) => {
-
-  const theme = useTheme()
-  const offSetPageIndex = table.getState().pagination.pageIndex + 1
-
   const { pageIndex, pageSize } = table.getState().pagination;
   const total = table.getPreFilteredRowModel().rows.length;
 
-  const start = total === 0 ? 0 : pageIndex * pageSize + 1;
-  const end = Math.min((pageIndex + 1) * pageSize, total);
-
   return (
-    <PaginationWrapper>
-      <StyledBody2SemiBold>
-        {`${start}-${end} of ${total}`}
-      </StyledBody2SemiBold>
-      <PaginationButtons>
-        <PaginationButton
-          disabled={!table.getCanPreviousPage()}
-          onClick={table.firstPage}
-        >
-          <FiChevronsLeft size={16} color={theme.colors.dark.darkGrey} />
-          <Body3>First</Body3>
-        </PaginationButton>
-        <PaginationButton
-          disabled={!table.getCanPreviousPage()}
-          onClick={table.previousPage}
-        >
-          <FiChevronLeft size={16} color={theme.colors.dark.darkGrey} />
-          <Body3>Back</Body3>
-        </PaginationButton>
-
-        <PaginationNavButton>
-          <StyledBody2Bold>{`${offSetPageIndex} `}</StyledBody2Bold>
-          <Body2>{`of ${table.getPageCount()}`}</Body2>
-        </PaginationNavButton>
-
-        <PaginationButton
-          disabled={!table.getCanNextPage()}
-          onClick={table.nextPage}
-        >
-          <Body3>Next</Body3>
-          <FiChevronRight size={16} color={theme.colors.dark.darkGrey} />
-        </PaginationButton>
-        <PaginationButton
-          disabled={!table.getCanNextPage()}
-          onClick={table.lastPage}
-        >
-          <Body3>Last</Body3>
-          <FiChevronsRight size={16} color={theme.colors.dark.darkGrey} />
-        </PaginationButton>
-      </PaginationButtons>
-    </PaginationWrapper>
+    <BasePagination
+      pageIndex={pageIndex}
+      pageCount={table.getPageCount()}
+      pageSize={pageSize}
+      total={total}
+      canPreviousPage={table.getCanPreviousPage()}
+      canNextPage={table.getCanNextPage()}
+      onFirstPage={table.firstPage}
+      onPreviousPage={table.previousPage}
+      onNextPage={table.nextPage}
+      onLastPage={table.lastPage}
+    />
   )
 }
-
-const StyledBody2Bold = styled(Body2SemiBold)`
-  color: ${props => props.theme.colors.dark.darkGrey};
-  font-weight: 700;
-`;
-
-const StyledBody2SemiBold = styled(Body2SemiBold)`
-  color: ${props => props.theme.colors.dark.darkGrey};
-`;
-
-const Body2 = styled(Body2Regular)`
-  color: ${props => props.theme.colors.dark.darkGrey};
-`;
-
-const PaginationWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 12px 0;
-  width: 100%;
-  margin: 0;
-`;
-
-const PaginationButton = styled.button<{
-  disabled: boolean
-}>`
-  all: unset;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  padding: 6px 10px;
-  color: ${props => props.theme.colors.dark.darkGrey};
-  border-radius: 4px;
-  border: 2px solid white;
-  background: white;
-  gap: 8px;
-
-  ${({ disabled }) => disabled && `
-    opacity: 0.5;
-    cursor: not-allowed;
-  `}
-  
-  ${({ disabled, theme }) => !disabled && `
-    &:hover {
-      background: ${darken(0.1, 'white')};
-      border-color: ${darken(0.1, 'white')};
-      color: ${theme.colors.dark.black};
-    }
-    &:active {
-      background: white;
-      border-color: ${theme.colors.green4};
-      color: ${theme.colors.dark.black};
-    }
-  `}
-`
-
-const PaginationButtons = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-`;
-
-const PaginationNavButton = styled(PaginationButtons)`
-  display: flex;
-  align-items: center;
-  gap: 4px;
-`;
